+++ src/connect/components/form/index.ts
@@ -1,1 +1,4 @@
+export * from './add-public-drive-form';
+export * from './create-local-drive-form';
 export * from './drive-settings-form';
+export * from './inputs';
